requirements.txt
annotated-types==0.7.0
anyio==4.10.0
# Editable Git install with no remote (bamboo==0.1.0)
-e /mnt/c/Users/tq/CodeProjects/LLM-EF
certifi==2025.8.3
# Editable Git install with no remote (df-llm==0.1.0)
-e /mnt/c/Users/tq/CodeProjects/LLM-EF
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
numpy==2.3.3
openai==1.108.0
pandas==2.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
